Module Osnap.Runner

Runner has tree mode:

Interactive:

Interactive mode provides an interactive runner, displaying differences between old and new snapshots. Promoting new version is proposed at every diff, exit when new diff is not promoted, considered as an error.

Example:

let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (+)

(* .osnap/add:

   f 5 6 = 11; *)

(* Then, the function under test is update *)
let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (fun _ _ -> 0)

let _ = Runner.(run_tests ~mode:Interactive [test])

(* .osnap/add
   -| f 5 6 = 11;
   +| f 5 6 = 0

   Promote changes ? [Y\n] *)

Promote:

Promote mode provides an automatic promotion of diffs.

Example:

let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (+)

(* .osnap/add:

   f 5 6 = 11; *)

(* Then, the function under test is update *)
let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (fun _ _ -> 0)

let _ = Runner.(run_tests ~mode:Promote [test])

(* .osnap/add has been promoted
   -| f 5 6 = 11;
   +| f 5 6 = 0
*)

Error:

Error mode will raise errors on diffs.

let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (+)

(* .osnap/add:

   f 5 6 = 11; *)

(* Then, the function under test is update *)
let test =
  let spec = OSnap.Spec.(int ^> int ^>> int) in
  Osnap.Test.make ~count:1 ~path:".osnap/add" ~spec (fun _ _ -> 0)

let _ = Runner.(run_tests ~mode:Promote [test])

(* .osnap/add has differences, exit
   -| f 5 6 = 11;
   +| f 5 6 = 0
*)

type mode =

| Interactive
Interactive mode, diffs will be displayed and ask for promote

| Promote
Promote mode, will promote every diff

| Error
Error mode, raises error on diff

val run_tests : ?⁠mode:mode -> Test.t list -> int

run_tests tests executes tests, default mode is Error.