Module Osnap__.Diff

type t =

| Same
old and new version are equals

| Diff of string
difference between old and new

| New of string
only new version

val pp : Stdlib.Format.formatter -> t -> unit

val diff : string option -> string -> t

diff prev next computes diff between prev and next using Patdiff If prev is absent, returns New [next]