Module Osnap__.Interpreter

type ('fn, 'r) args =

| Nil : ('r, 'r) args

| Cons : 'a * ('fn, 'r) args -> ('a -> 'fn, 'r) args

val spec_to_args : ('a, 'b) Osnap.Spec.t -> ('a, 'b) args

spec_to_args spec instantiate arguments values using generators inside spec

type _ expr =

| Apply : ('a -> 'b) expr * 'a expr -> 'b expr

| Term : 'a -> 'a expr

| Fun : ('a -> 'b) -> ('a -> 'b) expr

expr is an intern representation of a function application, Term inside expr are create using args.

val args_to_expr : 'a expr -> ('a, 'b) args -> 'b expr

args_to_expr f args create a function application representation using type expr. f must be a term Fun f, the expression will then be evaluated on that function f

val spec_to_expr : ('a -> 'b, 'c) Osnap.Spec.t -> ('a -> 'b) -> 'c expr

spec_to_expr spec creates an expr using spec_to_args and args_to_expr

val interpret : 'a expr -> 'a

interpret expr evaluates expr, providing the function a full aplication, returning its result.