Module Osnap

OCaml random snapshot testing

module Spec : sig ... end

Spec takes inspiration from OCaml's property-based testing library Monolith. The idea is that the programmer describes the function specification through the set of generators from QCheck and a printer.

module Test : sig ... end

module Runner : sig ... end