Module Osnap.Runner

type mode =

| Interactive
Interactive mode, diffs will be displayed and ask for promote

| Promote
Promote mode, will promote every diff

| Error
Error mode, raises error on diff

val run_tests : ?⁠mode:mode -> Test.t list -> int

run_tests tests runs suite of tests and print its results, default mode is Error.

returns

an error code, 0 if all tests passed, 1 otherwise