Module Osnap.Spec

include module type of Spec

module Gen = Spec.Gen

type 'a gen = 'a QCheck.Gen.t

'a gen is used to generate random values inside spec. QCheck combinators are available using Spec.Gen.

type 'a printer = 'a -> string

'a printer is used to store randomly generated values

type 'a spec = {

gen : 'a gen;

printer : 'a printer;

}

'a spec combines an 'a gen and a printer

type ('fn, 'r) t =

| Result : 'a printer -> ('a, 'a) t

| Arrow : 'a spec * ('fn, 'r) t -> ('a -> 'fn, 'r) t

t is the specification type, describing a function. Thus t declaration must end with (^>>).

val int : int spec

int specification

val list : 'a spec -> 'a list spec

list spec creates a list spec for spec

val (^>) : 'a spec -> ('b, 'c) t -> ('a -> 'b, 'c) t

(^>) x y combines spec x and y to create x -> y

val (^>>) : 'a spec -> 'b printer -> ('a -> 'b, 'b) t

(^>>) x res combines a specification and printer for the result type